chore(#14): move environment variable aware fastify instrumentation tests to separate file
