+++ test/env-vars.test-d.ts
@@ -1,0 +1,16 @@
+import { expectAssignable } from 'tsd'
+import { InstrumentationBase, InstrumentationConfig } from '@opentelemetry/instrumentation'
+import { fastify as Fastify } from 'fastify'
+
+import { FastifyOtelInstrumentation, FastifyOtelInstrumentationOpts } from '..'
+
+expectAssignable<InstrumentationBase>(new FastifyOtelInstrumentation())
+expectAssignable<InstrumentationConfig>({ servername: 'server', enabled: true } as FastifyOtelInstrumentationOpts)
+expectAssignable<InstrumentationConfig>({} as FastifyOtelInstrumentationOpts)
+
+const app = Fastify()
+app.register(new FastifyOtelInstrumentation().plugin())
+app.register((nested, _opts, done) => {
+  nested.register(new FastifyOtelInstrumentation().plugin())
+  done()
+})
